+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/General/Axes/XAxisTickFormattedOverflowValue.test.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import LineGraph from '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Line/LineGraph';
+import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
+import getGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/generalDefaultXAxisOverflow';
+
+const graphConfig = utils.deepClone(getGraphConfig('#xAxisTicksVlauesOverflow'));
+
+export default () => <LineGraph graphID="xAxisTicksVlauesOverflow" graphConfig={graphConfig} />;