+++ packages/carbon-graphs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added filter method for overflow tick values.
+
 * Changed
   * Updated d3 dependencies to only pull in the necessary modules.
   * Removed dependency on `d3-color@1` which has a known CVE issue.